obstack: Undefine the ELIDE_CODE macro explicitly
The obstack symbols can be exported by multiple GNU libraries, for
example by libc and libiberty. The ELIDE_CODE macro ensures that the
obstack symbols are not exported by multiple GNU libraries on the same
system.

In our case, we want to build this library in a GNU-based system, but we
want to use the library in a musl-based system. Thus, we need the
obstack symbols to be exported. To accomplish this we undefine the
ELIDE_CODE macro explicitly in the source code.

Signed-off-by: Nikos Dragazis <ndragazis@arrikto.com>